% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMEANS.R
\name{apply.kmeans}
\alias{apply.kmeans}
\title{Applies K-Means algorithm to the given data data.mat}
\usage{
apply.kmeans(data.mat, centers, iter.max = 10)
}
\arguments{
\item{data.mat}{Dataset of type matrix (List of double) and shape (n x d), where n is the number of observations and d is the dimension of the dataset.}

\item{centers}{The number of clusters.}

\item{iter.max}{Maximum number of iterations of Kmeans clustering algorithm.}
}
\value{
A class where "tot.withinss" is the within sum of squared errors, and "cluster" is the center of the clusters of shape (k x n).

  For description of how KMeans clustering algorithm works, please see the following url.
  \url{https://en.wikipedia.org/wiki/K-means_clustering}
}
\description{
\code{apply.kmeans} returns the within sum of squared distances (errors) from the cluster to the observations. It also provides the K centers of the clusters.
}
\examples{
apply.kmeans(iris[,-5], 3, 300)

\dontrun{
apply.kmeans(iris[,-5], 3, 300)
}

}
