% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{calculate_wcss}
\alias{calculate_wcss}
\title{Calculate WCSS for Different Cluster Counts}
\usage{
calculate_wcss(data, xvar, yvar, max_k = 10)
}
\arguments{
\item{data}{A data frame.}

\item{xvar}{X variable name.}

\item{yvar}{Y variable name.}

\item{max_k}{Max number of clusters to test.}
}
\value{
A vector of WCSS values.
}
\description{
Calculate WCSS for Different Cluster Counts
}
