% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_create.R
\name{cluser_create}
\alias{cluser_create}
\title{creates k means clusters}
\usage{
cluser_create(cols, cluster_num)
}
\arguments{
\item{cols}{The varibles of data uses}

\item{cluster_num}{Number of clusters}
}
\description{
Takes variables and number of clusters and uses that to create the k means clusters
}
