% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot.R
\name{cluster_plot}
\alias{cluster_plot}
\title{Plots kmeans clusters}
\usage{
cluster_plot(data, cluster_num, pc, ce)
}
\arguments{
\item{data}{The data frame or tibble used in the plot}

\item{cluster_num}{Number of clusters}

\item{pc}{Used in place of pch to plot}

\item{ce}{used in place of cex to plot}
}
\value{
output A description of the object the function outputs
}
\description{
Plots kmeans clusters
}
