\name{kmBlock-package}
\alias{kmBlock-package}
\alias{kmBlock}
\docType{package}
\title{
\packageTitle{kmBlock}
}
\description{
\packageDescription{kmBlock}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kmBlock}
\packageIndices{kmBlock}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{kmBlock}

Maintainer: \packageMaintainer{kmBlock}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
