% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_cluster.R
\name{kmeans_cluster}
\alias{kmeans_cluster}
\title{kmeans_cluster}
\usage{
kmeans_cluster(X, center, max.iter, tol)
}
\arguments{
\item{X}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{center}{either the number of clusters, say \(k\), or a set of initial (distinct) cluster centres. If a number, a random center will be assigned.}

\item{max.iter}{the maximum number of iterations allowed.}

\item{tol}{Parameters for determining convergence.}
}
\value{
Contains cell cluster labels for each cell in x and the center for each cluster
}
\description{
Perform k-means clustering on a data matrix.
}
\examples{


}
