% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_label_rematch.R
\name{kmeans_label_rematch}
\alias{kmeans_label_rematch}
\title{kmeans_label_rematch}
\usage{
kmeans_label_rematch(Y_train, Y_actual, Y_test = NULL, Y_test_actual = NULL)
}
\arguments{
\item{Y_train}{label got from the kmeans_cluster from training data}

\item{Y_actual}{the actual label}

\item{Y_test}{predict label for testing data (optional)}

\item{Y_test_actual}{the actual test label(optional)}
}
\value{
the vector of rematched label and training_accuracy.
   If Y_test_actual and Y_test are not null, return the testing accuracy as well.
}
\description{
match the true label based on the label existed
}
\examples{
X1 = c(1,3,2,2,1,2,2)
X2 = c(2,5,1,1,2,1,1)
kmeans_label_rematch(X1,X2)

}
