% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_shiny.R
\name{kmeans_shiny}
\alias{kmeans_shiny}
\title{Runs a shiny app of kemans for given data.}
\usage{
kmeans_shiny(data)
}
\arguments{
\item{data}{The data frame or tibble used in the shiny app.}
}
\description{
A more detailed description of what the function is and how
it works. It may be a paragraph that should not be separated
by any spaces.
}
\examples{
kmeans_shiny(iris)
}
