% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_best_k.R
\name{knn_best_k}
\alias{knn_best_k}
\title{gets best k value for knn of iris data}
\usage{
knn_best_k(varlist)
}
\arguments{
\item{varlist}{The list of variables}
}
\value{
Best k value for given varibles.
}
\description{
Take checked colums from the iris dataset and provieds the accuracy maximizing k
}
\examples{
knn_best_k(c('Sepal.Length','Sepal.Width','Petal.Length')) 
}
\keyword{keywords}
