% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_plot.R
\name{knn_plot}
\alias{knn_plot}
\title{Plots KNN of iris data}
\usage{
knn_plot(vlist, k)
}
\arguments{
\item{varlist}{The list of variables}
}
\value{
Graph of kmeans colored by species
}
\description{
Take checked colums from the iris dataset and plots the kmeans output with the top 2 selected on the x
and y axis
}
\examples{
knn_best_k(c('Sepal.Length','Sepal.Width','Petal.Length')) 
}
\keyword{keywords}
