% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{perform_kmeans}
\alias{perform_kmeans}
\title{Perform K-Means Clustering}
\usage{
perform_kmeans(data, xvar, yvar, centers = 3)
}
\arguments{
\item{data}{A data frame with at least two numeric columns.}

\item{xvar}{The name of the column to use as the x-axis.}

\item{yvar}{The name of the column to use as the y-axis.}

\item{centers}{The number of clusters.}
}
\value{
A kmeans object containing cluster assignments and centers.
}
\description{
Perform K-Means Clustering
}
