% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{Plot K-Means Clusters}
\usage{
plot_kmeans(data, kmeans_result, xvar, yvar)
}
\arguments{
\item{data}{A data frame.}

\item{kmeans_result}{A kmeans object.}

\item{xvar}{X variable name.}

\item{yvar}{Y variable name.}
}
\value{
A scatter plot.
}
\description{
Plot K-Means Clusters
}
