% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_cluster_predict.R
\name{kmeans_cluster_predict}
\alias{kmeans_cluster_predict}
\title{kmeans_cluster_predict}
\usage{
kmeans_cluster_predict(X, center)
}
\arguments{
\item{X}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{center}{the centres for each clusters}
}
\value{
label for each row observations
}
\description{
predict the label for each new data it belongs
}
\examples{


}
